+++ data/src/main/java/com/onmi/data/repository/MealRepositoryImpl.kt
@@ -1,13 +1,13 @@
 package com.onmi.data.repository
 
 import com.onmi.data.dto.meal.response.toModel
-import com.onmi.data.service.MealService
+import com.onmi.data.datasource.MealDataSource
 import com.onmi.domain.model.meal.response.GetMealsResponseModel
 import com.onmi.domain.repository.MealRepository
 import javax.inject.Inject
 
 class MealRepositoryImpl @Inject constructor(
-    private val service: MealService,
+    private val service: MealDataSource,
 ) : MealRepository {
 
     override suspend fun getMeals(
